# pkg/generated/models/pir_query.go
// Code generated by go-swagger; DO NOT EDIT.

//
// Copyright 2021 The Sigstore Authors.
//
// Licensed under the Apache License, Version 2.0 (the "License");
// you may not use this file except in compliance with the License.
// You may obtain a copy of the License at
//
//     http://www.apache.org/licenses/LICENSE-2.0
//
// Unless required by applicable law or agreed to in writing, software
// distributed under the License is distributed on an "AS IS" BASIS,
// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
// See the License for the specific language governing permissions and
// limitations under the License.
//

package models

import (
	"context"

	"github.com/go-openapi/errors"
	"github.com/go-openapi/strfmt"
	"github.com/go-openapi/swag/jsonutils"
	"github.com/go-openapi/validate"
)

// PirQuery pir query
//
// swagger:model PirQuery
type PirQuery struct {

	// public context
	// Required: true
	// Format: byte
	PublicContext *strfmt.Base64 `json:"publicContext"`

	// query blocks
	// Required: true
	QueryBlocks []strfmt.Base64 `json:"queryBlocks"`
}

// Validate validates this pir query
func (m *PirQuery) Validate(formats strfmt.Registry) error {
	var res []error

	if err := m.validatePublicContext(formats); err != nil {
		res = append(res, err)
	}

	if err := m.validateQueryBlocks(formats); err != nil {
		res = append(res, err)
	}

	if len(res) > 0 {
		return errors.CompositeValidationError(res...)
	}
	return nil
}

func (m *PirQuery) validatePublicContext(formats strfmt.Registry) error {

	if err := validate.Required("publicContext", "body", m.PublicContext); err != nil {
		return err
	}

	return nil
}

func (m *PirQuery) validateQueryBlocks(formats strfmt.Registry) error {

	if err := validate.Required("queryBlocks", "body", m.QueryBlocks); err != nil {
		return err
	}

	return nil
}

// ContextValidate validates this pir query based on context it is used
func (m *PirQuery) ContextValidate(_ context.Context, _ strfmt.Registry) error {
	return nil
}

// MarshalBinary interface implementation
func (m *PirQuery) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return jsonutils.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *PirQuery) UnmarshalBinary(b []byte) error {
	var res PirQuery
	if err := jsonutils.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}